refactored after clean up.
