+++ source/dotnet/wholesale-api/WebApi.IntegrationTests/Fixtures/TestCommon/Fixture/WebApi/WholesaleWebApiFixture.cs
@@ -89,17 +89,17 @@ namespace Energinet.DataHub.Wholesale.WebApi.IntegrationTests.Fixtures.TestCommo
             Environment.SetEnvironmentVariable(nameof(ServiceBusOptions.SERVICE_BUS_MANAGE_CONNECTION_STRING), ServiceBusResourceProvider.ConnectionString);
 
             // Add events configuration variables
-            Environment.SetEnvironmentVariable(nameof(Energinet.DataHub.Wholesale.Events.Application.Options.ServiceBusOptions.SERVICE_BUS_MANAGE_CONNECTION_STRING), ServiceBusResourceProvider.ConnectionString);
+            Environment.SetEnvironmentVariable(nameof(ServiceBusOptions.SERVICE_BUS_MANAGE_CONNECTION_STRING), ServiceBusResourceProvider.ConnectionString);
             await ServiceBusResourceProvider
                 .BuildQueue("sbq-wholesale-inbox")
-                .SetEnvironmentVariableToQueueName(nameof(Energinet.DataHub.Wholesale.Events.Application.Options.ServiceBusOptions.WHOLESALE_INBOX_MESSAGE_QUEUE_NAME))
+                .SetEnvironmentVariableToQueueName(nameof(ServiceBusOptions.WHOLESALE_INBOX_MESSAGE_QUEUE_NAME))
                 .CreateAsync();
 
             await ServiceBusResourceProvider
                 .BuildQueue("sbq-edi-inbox")
-                .SetEnvironmentVariableToQueueName(nameof(EdiInboxOptions.EDI_INBOX_MESSAGE_QUEUE_NAME))
+                .SetEnvironmentVariableToQueueName(nameof(ServiceBusOptions.EDI_INBOX_MESSAGE_QUEUE_NAME))
                 .CreateAsync();
-            Environment.SetEnvironmentVariable(nameof(EdiInboxOptions.EDI_INBOX_MESSAGE_QUEUE_NAME), ServiceBusResourceProvider.ConnectionString);
+            Environment.SetEnvironmentVariable(nameof(ServiceBusOptions.EDI_INBOX_MESSAGE_QUEUE_NAME), ServiceBusResourceProvider.ConnectionString);
 
             Environment.SetEnvironmentVariable(nameof(DateTimeOptions.TIME_ZONE), "Europe/Copenhagen");
