+++ source/dotnet/wholesale-api/WebApi.UnitTests/WebApi/Configuration/Options/OptionsTests.cs
@@ -30,16 +30,16 @@ public class OptionsTests
     [InlineAutoMoqData(typeof(AppInsightOptions), 1, "APPINSIGHTS_INSTRUMENTATIONKEY")]
     [InlineAutoMoqData(
         typeof(ServiceBusOptions),
-        4,
+        5,
         "SERVICE_BUS_SEND_CONNECTION_STRING",
         "SERVICE_BUS_MANAGE_CONNECTION_STRING",
         "INTEGRATIONEVENTS_TOPIC_NAME",
-        "WHOLESALE_INBOX_MESSAGE_QUEUE_NAME")]
+        "WHOLESALE_INBOX_MESSAGE_QUEUE_NAME",
+        "EDI_INBOX_MESSAGE_QUEUE_NAME")]
     [InlineAutoMoqData(typeof(DateTimeOptions), 1, "TIME_ZONE")]
     [InlineAutoMoqData(typeof(ConnectionStringsOptions), 1, "DB_CONNECTION_STRING")]
     [InlineAutoMoqData(typeof(DatabricksOptions), 3, "DATABRICKS_WORKSPACE_URL", "DATABRICKS_WORKSPACE_TOKEN", "DATABRICKS_WAREHOUSE_ID")]
     [InlineAutoMoqData(typeof(DeltaTableOptions), 2, "SCHEMA_NAME", "ENERGY_RESULTS_TABLE_NAME")]
-    [InlineAutoMoqData(typeof(EdiInboxOptions), 1, "EDI_INBOX_MESSAGE_QUEUE_NAME")]
     public void Options_HaveTheCorrectSettingNamesAndNumberOfSettings(Type sut, int settingsCount, params string[] expectedNames)
     {
         // Arrange & Act
